Move generic redis pool/cluster code to datastore/redis
