+++ server/datastore/redis/redis_test.go
@@ -1,10 +1,11 @@
-package pubsub
+package redis
 
 import (
 	"fmt"
 	"testing"
 	"time"
 
+	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/gomodule/redigo/redis"
 	"github.com/mna/redisc"
 	"github.com/stretchr/testify/require"
@@ -13,8 +14,8 @@ import (
 func TestEachRedisNode(t *testing.T) {
 	const prefix = "TestEachRedisNode:"
 
-	runTest := func(t *testing.T, store *redisQueryResults) {
-		conn := store.Pool().Get()
+	runTest := func(t *testing.T, pool fleet.RedisPool) {
+		conn := pool.Get()
 		defer conn.Close()
 		if rc, err := redisc.RetryConn(conn, 3, 100*time.Millisecond); err == nil {
 			conn = rc
@@ -26,7 +27,7 @@ func TestEachRedisNode(t *testing.T) {
 		}
 
 		var keys []string
-		err := EachRedisNode(store.Pool(), func(conn redis.Conn) error {
+		err := EachRedisNode(pool, func(conn redis.Conn) error {
 			var cursor int
 			for {
 				res, err := redis.Values(conn.Do("SCAN", cursor, "MATCH", prefix+"*"))
@@ -48,14 +49,47 @@ func TestEachRedisNode(t *testing.T) {
 	}
 
 	t.Run("standalone", func(t *testing.T) {
-		store, teardown := SetupRedisForTest(t, false)
+		pool, teardown := setupRedisForTest(t, false)
 		defer teardown()
-		runTest(t, store)
+		runTest(t, pool)
 	})
 
 	t.Run("cluster", func(t *testing.T) {
-		store, teardown := SetupRedisForTest(t, true)
+		pool, teardown := setupRedisForTest(t, true)
 		defer teardown()
-		runTest(t, store)
+		runTest(t, pool)
 	})
+}
+
+func setupRedisForTest(t *testing.T, cluster bool) (pool fleet.RedisPool, teardown func()) {
+	var (
+		addr     = "127.0.0.1:"
+		password = ""
+		database = 0
+		useTLS   = false
+		port     = "6379"
+	)
+	if cluster {
+		port = "7001"
+	}
+	addr += port
+
+	pool, err := NewRedisPool(addr, password, database, useTLS)
+	require.NoError(t, err)
+
+	conn := pool.Get()
+	defer conn.Close()
+	_, err = conn.Do("PING")
+	require.Nil(t, err)
+
+	teardown = func() {
+		err := EachRedisNode(pool, func(conn redis.Conn) error {
+			_, err := conn.Do("FLUSHDB")
+			return err
+		})
+		require.NoError(t, err)
+		pool.Close()
+	}
+
+	return pool, teardown
 }
